Add Rails Engine class; reorganize rails framework dir
