+++ lib/paper_trail/frameworks/rails/engine.rb
@@ -1,0 +1,6 @@
+module PaperTrail
+  module Rails
+    class Engine < ::Rails::Engine
+    end
+  end
+end
